swtchart: moving customized code into hpcviewer so that we can update/merge easily from the origin repository.
com.graphbuilder: similarly, we minimize customization as much as
possible. However, I haven't seen the original repository of this
library. 

Both swtchart and com.graphbuilder are now required plugin for
hpcviewer. This is to ease building hpcviewer RCP. Originally we have to
exporting them into jar and then importing into hpcviewer. Now they are
needed to be in the same workspace as hpcviewer.
I found this also a cleaner way to include external library.